// src/app/enums/top-customer-report-type.enum.ts
export enum TopCustomerReportTypeEnum {
    visitsPerMonth = 'visits per month',
    totalVisits = 'total visits',
    revenuePerVisit = 'revenue per visit',
    revenuePerMonth = 'revenue per month',
    totalRevenue = 'total revenue',
    tipsPerVisit = 'tips per visit',
    tipsPerMonth = 'tips per month',
    totalTips = 'total tips'
}